Split the signal_prepare test suite by stage
R2.2 of the transform cleanup plan: signal_prepare/tests.rs (32 tests)
becomes signal_prepare/tests/ grouped as staging, typing, recursion,
and verify with one shared fixture helper. Test names unchanged, 385
total, no widened visibility.
